同步最新的表结构

+++ yudao-admin-server/src/main/java/cn/iocoder/yudao/adminserver/modules/system/dal/mysql/user/SysUserMapper.java
@@ -68,7 +68,7 @@ public interface SysUserMapper extends BaseMapperX<SysUserDO> {
         return selectList("status", status);
     }
 
-    default List<SysUserDO> selectListByDeptIds(Collection<Integer> deptIds) {
+    default List<SysUserDO> selectListByDeptIds(Collection<Long> deptIds) {
         return selectList("dept_id", deptIds);
     }
